Change the signature of SignalBlock.process method
Traits with generic methods cannot be made into objects, so the method
signature for the process method had to be changed, in order to allow
for the SignalBlock trait to be made into an object. Sadly, this means
that the process method will be restricted to f32 types, but that should
be fine.